modify videoInput component so the user can only add either a video or a video link. Not both at the same time
