+++ zubhub_frontend/zubhub/src/components/form/videoInput/videoInput.styles.js
@@ -12,6 +12,7 @@ export const videoInputStyles = theme => ({
         flexDirection: 'column',
         gap: 10,
         cursor: 'pointer',
+        textAlign: 'center'
     },
     input: {
         visibility: 'hidden'
