+++ zubhub_frontend/zubhub/src/hooks/useDomElementHeight.hook.js
@@ -5,11 +5,16 @@ export function useDomElementHeight(elementID) {
 
     useEffect(() => {
         if (document && elementID) {
-            let nav = document.getElementById(elementID);
-            let height = nav.offsetHeight
-            setHeight(height)
+            let node = document.getElementById(elementID);
+            handleHeight(node)
+            window.addEventListener('resize', () => handleHeight(node))
         }
     }, [])
+
+    const handleHeight = (node) => {
+        let height = node.offsetHeight
+        setHeight(height)
+    }
 
     return { height }
 }
